Db model upgrade

+++ app/main.py
@@ -1,39 +1,76 @@
-from flask import Flask, request, render_template, redirect
-from auth import auth_bp, login_required
+from flask import Flask, request, jsonify, Response
+
+from models import User, SkillUser, SkillName, Experience
+from session import get_session
 
 import json
-
-from models import Cv
-from session import get_session
 
 session = get_session(echo=False)
 
 app = Flask(__name__)
-app.secret_key = 'tajny-klucz-9523'
-app.register_blueprint(auth_bp)
+app.config['JSON_SORT_KEYS'] = False
 
 
 @app.route('/')
 def index():
-    return "For API please use /api/cv or /api/cv/<id>"  # TODO jak to zmusić do wyświetlania <id> w safari?
+    msg = "For API please use /api/cv or /api/cv/<id>"
+    return Response(msg, mimetype='text/plain')
 
 
 @app.route('/api/cv', methods=['GET'])
-@app.route('/api/cv/', methods=['GET'])  # TODO czy tak się robi? czy w <id> też dorobić / po adresie?
+@app.route('/api/cv/', methods=['GET'])
 def api_cv_get():
+    all_db_records = [user.object_as_dict() for user in session.query(User)]
 
-    all_db_records = [cv_instance.object_as_dict() for cv_instance in session.query(Cv)]
-
-    return json.dumps(all_db_records)
+    return jsonify(all_db_records)
 
 
 @app.route('/api/cv', methods=['POST'])
 @app.route('/api/cv/', methods=['POST'])
 def api_cv_post():
 
-    query_data = request.get_json()
+    # ADD NEW CV BASED ON JSON DATA
 
-    session.add(Cv(**query_data))
+    json_data = request.get_json()
+
+    # create new_cv object and map basic user data from json:
+    new_cv = User()
+
+    if "username" in json_data:
+        new_cv.username = json_data["username"]
+    if "password" in json_data:
+        new_cv.password = json_data["password"]
+
+    new_cv.firstname = json_data["firstname"]
+    new_cv.lastname = json_data["lastname"]
+
+    # add skills from json to new_cv object, if skills were provided:
+    for skill in json_data.get('skills', []):
+
+        # check if skill_name already exists
+        skill_name_obj = session.query(SkillName).filter_by(skill_name=skill["skill_name"]).first()
+
+        if not skill_name_obj:
+            skill_name_obj = SkillName(skill_name=skill["skill_name"])
+            session.add(skill_name_obj)
+
+        skill_object = SkillUser()
+        skill_object.skill = skill_name_obj
+        skill_object.skill_level = skill["skill_level"]
+
+        new_cv.skills.append(skill_object)
+
+    # add experience from json to new_cv object, if experience was provided:
+    for exp in json_data.get('experience', []):
+
+        exp_object = Experience()
+        exp_object.company = exp["company"]
+        exp_object.project = exp["project"]
+        exp_object.duration = exp["duration"]
+
+        new_cv.experience.append(exp_object)
+
+    session.add(new_cv)
     session.commit()
 
     return "", 201
@@ -41,8 +78,8 @@ def api_cv_post():
 
 @app.route('/api/cv/<id>', methods=['GET'])
 def api_cv_id_get(id=None):
+    one_db_record = session.query(User).get(id)
 
-    one_db_record = session.query(Cv).get(id)
     try:
         response = json.dumps(one_db_record.object_as_dict())
         return response
@@ -53,34 +90,73 @@ def api_cv_id_get(id=None):
 
 @app.route('/api/cv/<id>', methods=['PUT'])
 def api_cv_id_put(id=None):
+
+    # UPDATE CV OF ID [id] WITH JSON DATA
+
     json_data = request.get_json()
 
-    cv_being_updated = session.query(Cv).get(id)
+    cv_being_updated = session.query(User).get(id)
 
-    cv_being_updated.firstname = json_data["firstname"]
-    cv_being_updated.lastname = json_data["lastname"]
-    cv_being_updated.python = json_data["python"]
-    cv_being_updated.javascript = json_data["javascript"]
-    cv_being_updated.sql = json_data["sql"]
-    cv_being_updated.english = json_data["english"]
+    if cv_being_updated is None:
+        return "", 404
 
-    # TODO od PD - uprościć update objektu tak o:
-    # for key, value in json_data.items():
-    #     setattr(cv_being_updated, key, value)
+    print(cv_being_updated)
+
+    # update basic user's attributes:
+    for key, value in json_data.items():
+        if key not in ["skills", "experience"]:
+            setattr(cv_being_updated, key, value)
+
+    # overwrite user's skills (will be redesigned in future):
+    if "skills" in json_data.keys():
+
+        cv_being_updated.skills = []
+
+        for json_skill in json_data.get('skills', []):
+
+            skill_name_obj = session.query(SkillName).filter_by(skill_name=json_skill["skill_name"]).first()
+
+            if not skill_name_obj:
+                skill_name_obj = SkillName(skill_name=json_skill["skill_name"])
+                session.add(skill_name_obj)
+
+            skill_object = SkillUser()
+            skill_object.skill = skill_name_obj
+            skill_object.skill_level = json_skill["skill_level"]
+
+            cv_being_updated.skills.append(skill_object)
+
+    # overwrite user's experience (will be redesigned in future):
+    if "experience" in json_data.keys():
+
+        cv_being_updated.experience = []
+
+        for json_exp in json_data.get('experience', []):
+            exp_object = Experience()
+            exp_object.company = json_exp["company"]
+            exp_object.project = json_exp["project"]
+            exp_object.duration = json_exp["duration"]
+
+            cv_being_updated.experience.append(exp_object)
 
     session.commit()
 
-    return "", 202
+    return "", 200
 
 
 @app.route('/api/cv/<id>', methods=['DELETE'])
 def api_cv_id_delete(id=None):
 
-    cv_to_be_deleted = session.query(Cv).get(id)
-    session.delete(cv_to_be_deleted)
-    session.commit()
+    cv_to_be_deleted = session.query(User).get(id)
 
-    return "", 202
+    if cv_to_be_deleted is not None:
+        session.delete(cv_to_be_deleted)
+        session.commit()
+
+        return "", 204
+
+    else:
+        return "", 404
 
 
 if __name__ == '__main__':
